chessgame version 2.0
we separeted the pawn's concept in an abstract class : a piece. there
are same attribute and methode -> color,getColor,toString,a list of
movement where the piece can go

so create all the piece which compose a Chessgame : pawn, rook, bishop,
knight, queen and king.

two enum was created : one for the different color that a piece can
take(black or withe)
					   the second for the different direction that a piece can
go(up,down,right or left)
					   
					   
start to work on differents possibilities piece's movement: random,
buffered reader.